// nodes/HaloPSA/actions/interfaces/projects.ts
export interface HaloProjectCreate extends IDataObject {
	summary: string;
	
	dateoccurred?: string;
	details?: string;
	details_html?: string;
	tickettype_id?: number;
	status_id?: number;
	priority_id?: number;
	sla_id?: number;
	
	client_id?: number;
	site_id?: number;
	user_id?: number;
	
	agent_id?: number;
	team_id?: number;
	
	category_1?: string;
	category_2?: string;
	category_3?: string;
	category_4?: string;
	categoryid_1?: number;
	categoryid_2?: number;
	categoryid_3?: number;
	categoryid_4?: number;
	
	estimate?: number;
	estimatedays?: number;
	
	reportedby?: string;
	user_email?: string;
	emailtolist?: string;
	emailcclist?: string;
	phonenumber?: string;
	
	product_id?: number;
	matched_kb_id?: number;
	
	impact?: number;
	urgency?: number;
	impactlevel?: number;
	
	oppcontactname?: string;
	oppcompanyname?: string;
	oppemailaddress?: string;
	opptel?: string;
	oppaddr1?: string;
	oppaddr2?: string;
	oppaddr3?: string;
	oppaddr4?: string;
	opppostcode?: string;
	oppcountry?: string;
	oppregion?: string;
	opptype?: string;
	oppvalueadjusted?: number;
	opptimezonename?: string;
	oppdontaddtomailinglist?: boolean;
	
	cost?: number;
	quantity?: number;
	
	starttime?: string;
	starttimeslot?: number;
	targetdate?: string;
	targettime?: string;
	targettimeslot?: number;
	deadlinedate?: string;
	
	source?: number;
	supplier_status?: number;
	supplier_contract_id?: number;
	
	flagged?: boolean;
	onhold?: boolean;
	excludefromsla?: boolean;
	showforusers?: boolean;
	inactive?: boolean;
	is_vip?: boolean;
	isimportantcontact?: boolean;
	surveysent?: boolean;
	surveyneeded?: boolean;
	knownerror?: boolean;
	development?: boolean;
	reviewed?: boolean;
	
	organisation_id?: number;
	department_id?: number;
	team_department_id?: number;
	top_level_id?: number;
	
	workflow_id?: number;
	workflow_step?: number;
	workflow_seq?: number;
	pipeline_stage_id?: number;
	
	release_id?: number;
	release2_id?: number;
	release3_id?: number;
	release_important?: boolean;
	releasenotegroup_id?: number;
	
	alsoinform?: string;
	backoutplan?: string;
	communicationplan?: string;
	testplan?: string;
	riskdescription?: string;
	risklevel?: number;
	impactdescription?: string;
	changeinformation_html?: string;
	
	appointment_type?: number;
	resourcetype_id?: number;
	resource_booking_type?: number;
	
	projectinternaltask?: boolean;
	isbillable?: boolean;
	itemsarebillable?: boolean;
	actisbillabledefault?: boolean;
	invoiceseperatelyoverride?: boolean;
	purchaseordernumber?: string;
	budgettype_id?: number;
	
	component_id?: number;
	version_id?: number;
	contract_id?: number;
	
	pagerdutyid?: string;
	servicenow_id?: string;
	connectwise_id?: number;
	autotask_id?: number;
	ninja_alert_id?: string;
	ninja_device_id?: number;
	atera_alert_id?: number;
	
	site_timezone?: string;
	timezonename?: string;
	section_timezone?: string;
	
	inform3rdpartysystem?: boolean;
	forwardinboundupdates?: boolean;
	loggedoutofhdworkinghours?: boolean;
	sendprintrequest?: number;
	deadlinenotificationhours?: number;
	showonroadmap?: boolean;
	
	delivery_address?: HaloAddress;
	billing_address?: HaloAddress;
	
	ticket_tags?: string;
	
	is_sensitive?: boolean;
	hide_feedback?: boolean;
	
	sms_override?: string;
	
	itil_requesttype_id?: number;
	
	guid?: string;
	takenby?: string;
	loggedonbehalfby?: string;
	who?: string;
	article_type?: number;
	phonenumberfrom?: string;
	addressfrom?: string;
}

export interface HaloProjectUpdate extends Partial<HaloProjectCreate> {
	id: number;
}